+++ libblocky/src/instance/models/file.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct File {
     pub sha1: String,
-    pub size: i64,
+    pub size: usize,
     pub url: String,
 }
